+++ lib/commands/new.dart
@@ -3,7 +3,7 @@ import 'dart:io';
 
 import 'package:arceus/arceus.dart';
 import 'package:arceus/extensions.dart';
-import 'package:arceus/serekit/serekit.dart';
+import 'package:arceus/serekit/skit.dart';
 import 'package:arceus/version_control/constellation.dart';
 import 'package:args/command_runner.dart';
 import 'package:cli_spin/cli_spin.dart';
