+++ lib/commands/trim.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
 
 import 'package:arceus/extensions.dart';
-import 'package:arceus/serekit/serekit.dart';
+import 'package:arceus/serekit/skit.dart';
 import 'package:arceus/version_control/constellation.dart';
 import 'package:args/command_runner.dart';
 import 'package:arceus/arceus.dart';
